# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/admin/audit-log:
    get:
      operationId: listAuditLog
      tags: [admin]
      summary: List audit log entries (admin-only)
      x-status: stub
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: eventId
          schema: { $ref: '#/components/schemas/Uuid' }
        - in: query
          name: actorUserId
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: Paginated audit entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLog'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/admin/audit-log.csv:
    get:
      operationId: exportAuditLogCsv
      tags: [admin]
      summary: Export audit log as CSV (admin-only)
      x-status: stub
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: eventId
          schema: { $ref: '#/components/schemas/Uuid' }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: CSV stream.
          content:
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
        description: Opaque cursor returned by a prior page.
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...
    PaginatedAuditLog:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuditLogEntry' }
        nextCursor: { type: string, nullable: true }
...
    AuditActorKind:
      type: string
      enum: [user, system, cron, admin, webhook]
    AuditLogEntry:
      type: object
      required: [id, actorKind, action, createdAt]
      properties:
        id:
          type: string
          description: Bigint id encoded as string.
        actorUserId: { $ref: '#/components/schemas/Uuid' }
        actorKind: { $ref: '#/components/schemas/AuditActorKind' }
        action: { type: string }
        targetKind: { type: string, nullable: true }
        targetId: { type: string, nullable: true }
        eventId: { $ref: '#/components/schemas/Uuid' }
        ipHash: { type: string, nullable: true }
        userAgent: { type: string, nullable: true }
        payloadJsonb:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }